Annotations restructured to controllers and models annotations folders
